@@ -1,12 +1,20 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 24
-    buildToolsVersion "25.0.3"
+    signingConfigs {
+        debug {
+            storeFile file('C:\\Users\\Administrator\\android\\client\\key.jks')
+            keyAlias 'sj'
+            storePassword '596030631'
+            keyPassword '596030631'
+        }
+    }
+    compileSdkVersion 30
+    buildToolsVersion '30.0.3'
 
     defaultConfig {
-        minSdkVersion 15
-        targetSdkVersion 22
+        minSdkVersion 21
+        targetSdkVersion 30
         versionCode 1000
         versionName "1.0.0"
 
@@ -19,6 +27,7 @@ android {
         ndk {
             abiFilters "arm64-v8a", "armeabi-v7a"
         }
+        signingConfig signingConfigs.debug
 
     }
     buildTypes {
@@ -41,6 +50,11 @@ android {
             assets.srcDirs = ['src/main/assets', '../../../model']
         }
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+    ndkVersion '21.4.7075529'
 }
 
 task deleteFiles(type: Delete) {
@@ -66,7 +80,7 @@ preBuild.dependsOn(copyFiles)
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support:appcompat-v7:24.2.1'
+    implementation 'androidx.appcompat:appcompat:1.3.0'
 }
 
 
